+++ src/main/java/org/sharetrace/experiment/config/BaseParamsExperimentConfig.java
@@ -3,7 +3,7 @@ package org.sharetrace.experiment.config;
 import java.util.Optional;
 import org.immutables.value.Value;
 import org.sharetrace.experiment.GraphType;
-import org.sharetrace.util.range.DoubleRange;
+import org.sharetrace.util.range.FloatRange;
 import org.sharetrace.util.range.Range;
 
 @Value.Immutable
@@ -14,12 +14,12 @@ interface BaseParamsExperimentConfig {
   Optional<Integer> numNodes();
 
   @Value.Default
-  default Range<Double> transRates() {
-    return DoubleRange.of(0.1, 1.0, 0.1);
+  default Range<Float> transRates() {
+    return FloatRange.of(0.1, 1.0, 0.1);
   }
 
   @Value.Default
-  default Range<Double> sendCoeffs() {
-    return DoubleRange.of(0.1, 1.1, 0.1);
+  default Range<Float> sendCoeffs() {
+    return FloatRange.of(0.1, 1.1, 0.1);
   }
 }
