+++ src/main/java/org/sharetrace/model/BaseMsgParams.java
@@ -17,16 +17,16 @@ import org.sharetrace.util.Checks;
 @Value.Immutable(copy = true)
 abstract class BaseMsgParams {
 
-  public static final double MIN_TRANS_RATE = 0d;
-  public static final double MAX_TRANS_RATE = 1d;
-  public static final double MIN_SEND_COEFF = 0d;
-  public static final double MIN_TOLERANCE = 0d;
+  public static final float MIN_TRANS_RATE = 0f;
+  public static final float MAX_TRANS_RATE = 1f;
+  public static final float MIN_SEND_COEFF = 0f;
+  public static final float MIN_TOLERANCE = 0f;
 
   /**
    * Returns the rate at which the value of a {@link RiskScore} exponentially decreases as it
    * propagates from the source {@link User} during {@link RiskPropagation}.
    */
-  public abstract double transRate();
+  public abstract float transRate();
 
   /**
    * Returns the multiplier used to set the threshold of a {@link User} that determines if a
@@ -38,7 +38,7 @@ abstract class BaseMsgParams {
    * Given a positive transmission rate, a positive coefficient guarantees that {@link
    * RiskPropagation} terminates.
    */
-  public abstract double sendCoeff();
+  public abstract float sendCoeff();
 
   /**
    * Returns the extent to which a {@link RiskScore} is relevant after a contact occurs. When
@@ -70,7 +70,7 @@ abstract class BaseMsgParams {
    * Returns the maximum absolute difference between {@link RiskScore} values for them to be
    * approximately equal.
    */
-  public abstract double tolerance();
+  public abstract float tolerance();
 
   @Value.Check
   protected void check() {
